Major Update
At this point in time we are able to create a MusicNote object, the sound now plays well, the timer now works well, the lines are displayed. based on this we can now create different notes with if statements, we can plot the notes on the lines and create a list of music notes that we can later play back.